+++ server_variables.php
@@ -9,18 +9,18 @@
 
 namespace PMA;
 
-use PMA\libraries\controllers\ServerVariablesController;
+use PMA\libraries\controllers\server\ServerVariablesController;
 use PMA\libraries\Response;
 
 require_once 'libraries/common.inc.php';
 
 $container = libraries\di\Container::getDefaultContainer();
 $container->factory(
-    'PMA\libraries\controllers\ServerVariablesController'
+    'PMA\libraries\controllers\server\ServerVariablesController'
 );
 $container->alias(
     'ServerVariablesController',
-    'PMA\libraries\controllers\ServerVariablesController'
+    'PMA\libraries\controllers\server\ServerVariablesController'
 );
 $container->set('PMA\libraries\Response', Response::getInstance());
 $container->alias('response', 'PMA\libraries\Response');
